requirements.txt
openpyxl>=3.1
python-docx>=1.0
pyyaml>=6.0
jsonschema>=4.0
pytest>=7.0
